// src/interface/customer.interface.ts
 
export interface Customer {
    Customer_Name: string;
    Company: string;
    Phone_Number: string;
    Email: string;
    Country: string;
    Status: string;
  }
  export interface MenuItem {
    icon: string;
    title: string;
    altText: string;
    showChevron?: boolean;
    isActive:boolean
  }
  